Harden the RAVV size() bound and trim a redundant re-vectorize
Follow-up refinements to the sparse-ordinal fix, from review.

Fail fast instead of overflowing in RandomAccessVectorValues.size(). The ordinal upper bound highestUsedId() + 1 was cast to int with a plain (int) cast, which would silently wrap to a negative size() if the id space ever approached Integer.MAX_VALUE. Use Math.toIntExact at all three sites (EntityBackedVectorValues.size() and the two computed sizeSupplier lambdas in createVectorValues()/createCachingVectorValues()) so an oversized id space throws a clear ArithmeticException rather than returning a corrupt bound — consistent with jvector's int ordinal ceiling.

Avoid re-vectorizing the replaced entity on the common update path. In embedded mode internalUpdate computed contentChanged from the old vector unconditionally, re-running the vectorizer on every update even when the new vector is non-null. A non-null new vector is always a change (null->vec or vec->vec); only a null new vector needs the old vector, to distinguish vec->null (a change) from null->null (the only no-op). Compute the old vector solely on that path: contentChanged = vector != null || (replacedEntity != null && vectorize(replacedEntity) != null).

Correct the computedIdIndex field comment. It claimed the map is "written under the vector store's own monitor", but only the volatile field's one-time publication (the lazy build in computedIdIndex()) holds that monitor; per-entry put/remove by mutations rely on the ConcurrentHashMap's own thread-safety, serialized against each other by the parent GigaMap monitor the mutation already holds. Reworded to match.

Only behavioral change is skipping the redundant vectorize; the rest are a fail-fast guard and a comment. Verified across the full jvector fast and slow suites.

+++ gigamap/jvector/src/main/java/org/eclipse/store/gigamap/jvector/EntityBackedVectorValues.java
@@ -54,8 +54,9 @@ class EntityBackedVectorValues<E> implements RandomAccessVectorValues
         // count: the graph ordinal is the entity id, so deletion holes and null-embedding entities
         // (still counted in entityMap.size()) push the highest ordinal beyond the count. Matching the
         // id space keeps PQ encodeAll() (which builds a dense array of length size() and is indexed by
-        // ordinal) from skipping/overflowing high ordinals.
-        return (int)(this.entityMap.highestUsedId() + 1);
+        // ordinal) from skipping/overflowing high ordinals. toIntExact fails fast rather than silently
+        // overflowing to a negative size() near the int ordinal ceiling.
+        return Math.toIntExact(this.entityMap.highestUsedId() + 1);
     }
 
     @Override
